refactor: rename archiveSpace store
